# Only coerce time when comparing if necessary