@aaronjensen 2015-04-01
No problem. Thanks for the quick merge. Do you need another pull request for a back-port to 4.2 or is that automatic?