@rafaelfranca 2015-04-01
We usually don't backport performance improvements unless it is a regression, so I'd not backport this commit.